fix service delete logic, fix LensController to redirect to the correct route and fix form action URL in the camera edit

+++ src/main/java/com/pugliatti/andrea/retrofocus/service/CameraService.java
@@ -53,6 +53,8 @@ public class CameraService {
     }
 
     public void deleteById(Integer id) {
-        cameraRepository.deleteById(id);
+        if (cameraRepository.existsById(id)) {
+            cameraRepository.deleteById(id);
+        }
     }
 }
