+++ src/main/java/com/pugliatti/andrea/retrofocus/service/LensService.java
@@ -53,6 +53,8 @@ public class LensService {
     }
 
     public void deleteById(Integer id) {
-        lensRepository.deleteById(id);
+        if (lensRepository.existsById(id)) {
+            lensRepository.deleteById(id);
+        }
     }
 }
